move partnerIntegrations to strict registration
