+++ src/background/messenger/strict/registration.ts
@@ -52,6 +52,10 @@ import { locator, refreshServices } from "@/background/locator";
 import { closeTab, focusTab, openTab } from "@/background/tabs";
 import launchInteractiveOAuth2Flow from "@/background/auth/launchInteractiveOAuth2Flow";
 import { performConfiguredRequest } from "@/background/requests";
+import {
+  getPartnerPrincipals,
+  launchAuthIntegration,
+} from "@/background/partnerIntegrations";
 
 expectContext("background");
 
@@ -100,6 +104,9 @@ declare global {
     LAUNCH_INTERACTIVE_OAUTH_FLOW: typeof launchInteractiveOAuth2Flow;
 
     CONFIGURED_REQUEST: typeof performConfiguredRequest;
+
+    GET_PARTNER_PRINCIPALS: typeof getPartnerPrincipals;
+    LAUNCH_AUTH_INTEGRATION: typeof launchAuthIntegration;
   }
 }
 
@@ -148,5 +155,8 @@ export default function registerMessenger(): void {
     LAUNCH_INTERACTIVE_OAUTH_FLOW: launchInteractiveOAuth2Flow,
 
     CONFIGURED_REQUEST: performConfiguredRequest,
+
+    GET_PARTNER_PRINCIPALS: getPartnerPrincipals,
+    LAUNCH_AUTH_INTEGRATION: launchAuthIntegration,
   });
 }
