+++ src/extensionConsole/pages/onboarding/partner/ControlRoomOAuthForm.tsx
@@ -17,7 +17,7 @@
 
 import React, { useCallback, useContext } from "react";
 import { uuidv4 } from "@/types/helpers";
-import { launchAuthIntegration } from "@/background/messenger/api";
+import { launchAuthIntegration } from "@/background/messenger/strict/api";
 import Form, {
   type RenderBody,
   type RenderSubmit,
